packages/training/pyproject.toml
[project]
name = "training"
version = "0.1.0"
description = "Fine-tuning pipeline for PT-BR news summarization"
requires-python = ">=3.11,<3.13"
dependencies = [
  "accelerate>=0.27.2,<0.32.0",
  "datasets>=2.18.0,<2.20.0",
  "evaluate>=0.4.1,<0.5.0",
  "numpy>=1.26.0,<2",
  "peft>=0.9.0,<0.11.0",
  "pyyaml>=6.0.2",
  "rouge-score>=0.1.2",
  "safetensors>=0.4.2",
  "setuptools>=69.0.0",
  "torch==2.1.2",
  "transformers>=4.38.2,<4.40.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/training"]
